feat: :sparkles: add fluter modular
