+++ lib/main.dart
@@ -1,6 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+import 'package:glpi_hoiol/app/app_module.dart';
 import 'package:glpi_hoiol/app/app_widget.dart';
 
 void main() {
-  runApp(const AppWidget());
+  runApp(ModularApp(
+    module: AppModule(),
+    child: const AppWidget(),
+  ));
 }
