+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   dio: ^4.0.6
   uno: ^1.0.1+3
   bloc: ^8.1.0
+  flutter_modular: ^5.0.3
 
 dev_dependencies:
   flutter_test:
